deterministic_parameter_search/requirements.txt
antimony==2.12.0.2
appdirs==1.4.4
attrs==21.2.0
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
h5py==3.6.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lhsmdu==1.1
libroadrunner==2.0.5
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mpmath==1.2.1
networkx==2.5.1
numpy==1.19.3
packaging==20.9
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
phrasedml==1.1.1
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-libcombine==0.2.11
python-libnuml==1.1.4
python-libsbml==5.19.0
python-libsedml==2.0.20
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
retrying==1.3.3
rrplugins==2.0.4
sbml2matlab==1.2.3.1
scipy==1.6.3
six==1.16.0
SobolSequence==0.2
sympy==1.8
tellurium==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
